move model package
