+++ app/src/main/java/jp/chau2chaun2/honkot/sample/multimodule/di/AppModule.kt
@@ -3,19 +3,13 @@ package jp.chau2chaun2.honkot.sample.multimodule.di
 import android.content.Context
 import android.content.res.Resources
 import com.github.gfx.android.orma.AccessThreadConstraint
-import com.google.gson.FieldNamingPolicy
-import com.google.gson.Gson
-import com.google.gson.GsonBuilder
 import dagger.Module
 import dagger.Provides
 import jp.chau2chaun2.honkot.sample.multimodule.CustomApplication
 import jp.chau2chaun2.honkot.sample.multimodule.data.api.IQiitaService
+import jp.chau2chaun2.honkot.sample.multimodule.data.api_builder.ApiBuilder
+import jp.chau2chaun2.honkot.sample.multimodule.data.api_builder.CustomGsonBuilder
 import jp.chau2chaun2.honkot.sample.multimodule.model.OrmaDatabase
-import jp.chau2chaun2.honkot.sample.multimodule.util.DateTimeUtil
-import okhttp3.OkHttpClient
-import org.threeten.bp.ZonedDateTime
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
 import javax.inject.Singleton
 
 @Module
@@ -38,24 +32,7 @@ class AppModule {
 
     @Singleton
     @Provides
-    fun provideGson(): Gson {
-        return GsonBuilder()
-            .serializeNulls()
-            .setFieldNamingPolicy(FieldNamingPolicy.LOWER_CASE_WITH_UNDERSCORES)
-            .registerTypeAdapter(ZonedDateTime::class.java, DateTimeUtil.DateTimestampSerializer())
-            .registerTypeAdapter(ZonedDateTime::class.java, DateTimeUtil.DateTimestampDeserializer())
-            .create()
-    }
-
-    @Singleton
-    @Provides
-    fun provideQiitaService(gson: Gson): IQiitaService {
-        val okHttpClient = OkHttpClient.Builder().build()
-        val retrofit = Retrofit.Builder()
-            .baseUrl("https://qiita.com/")
-            .addConverterFactory(GsonConverterFactory.create(gson))
-            .client(okHttpClient)
-            .build()
-        return retrofit.create(IQiitaService::class.java)
+    fun provideQiitaService(customGsonBuilder: CustomGsonBuilder): IQiitaService {
+        return ApiBuilder().buildQiitaService("https://qiita.com/", customGsonBuilder.buildGson())
     }
 }
