+++ app/src/main/java/jp/chau2chaun2/honkot/sample/multimodule/vm/ListDatabaseFragmentViewModel.kt
@@ -4,15 +4,14 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import jp.chau2chaun2.honkot.sample.multimodule.model.Repo
+import jp.chau2chaun2.honkot.sample.multimodule.data.model.Repo
 import jp.chau2chaun2.honkot.sample.multimodule.repository.RepoRepository
-import jp.chau2chaun2.honkot.sample.multimodule.util.IDataLoading
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 class ListDatabaseFragmentViewModel @Inject constructor(
     private val repoRepository: RepoRepository
-) : IDataLoading, ViewModel() {
+) : jp.chau2chaun2.honkot.sample.multimodule.util.IDataLoading, ViewModel() {
 
     private val _loading = MutableLiveData<Boolean>()
