+++ data/model/src/main/java/jp/chau2chaun2/honkot/sample/multimodule/data/model/Repo.kt
@@ -1,9 +1,10 @@
-package jp.chau2chaun2.honkot.sample.multimodule.model
+package jp.chau2chaun2.honkot.sample.multimodule.data.model
 
 import android.provider.BaseColumns
 import com.github.gfx.android.orma.annotation.Column
 import com.github.gfx.android.orma.annotation.PrimaryKey
 import com.github.gfx.android.orma.annotation.Table
+import jp.chau2chaun2.honkot.sample.multimodule.data.model.ColorType
 
 @Table("repos")
 class Repo {
